+++ sdk/src/client/warp-gateway.js
@@ -1,5 +1,6 @@
 import { Rejected, fromPromise, of } from 'hyper-async'
 import { append, evolve, filter, pipe } from 'ramda'
+import { z } from 'zod'
 
 /**
  * Inject dependencies, so they are configurable,
@@ -28,17 +29,7 @@ export function deployContractWith ({ fetch, WARP_GATEWAY_URL, logger: _logger, 
 
   return (args) => {
     return of(args)
-      .map(logger.tap('deploying bundled contract via the warp gateway %s', `${WARP_GATEWAY_URL}/gateway/v2/contracts/deploy`))
-      /**
-       * The Warp Gateway requires that the Content-Type tag
-       * is a 'application/x.arweave-manifest+json'
-       *
-       * This is a requirement of the Warp Gateway, not SmartWeave,
-       * which is why this code exists here.
-       *
-       * Set 'data' in the data item to a valid, but empty, path manifest
-       * and replace the Content-Type flag to be the value Warp Gateway expects
-       */
+      .map(logger.tap('deploying bundled contract via the warp gateway %s', `${WARP_GATEWAY_URL}/gateway/contracts/deploy`))
       .map(evolve({ tags: pipe(replaceSdkTag, addNonceTag) }))
       /**
        * Sign with the provided signer
@@ -51,18 +42,18 @@ export function deployContractWith ({ fetch, WARP_GATEWAY_URL, logger: _logger, 
         of(signedDataItem)
           .chain(fromPromise(async (signedDataItem) =>
             fetch(
-            `${WARP_GATEWAY_URL}/gateway/v2/contracts/deploy`,
-            {
-              method: 'POST',
-              headers: {
-                'Accept-Encoding': 'gzip, deflate, br',
-                'Content-Type': 'application/json',
-                Accept: 'application/json'
-              },
-              body: JSON.stringify({
-                contract: signedDataItem.raw
-              })
-            }
+              `${WARP_GATEWAY_URL}/gateway/v2/contracts/deploy`,
+              {
+                method: 'POST',
+                headers: {
+                  'Accept-Encoding': 'gzip, deflate, br',
+                  'Content-Type': 'application/json',
+                  Accept: 'application/json'
+                },
+                body: JSON.stringify({
+                  contract: signedDataItem.raw
+                })
+              }
             )
           ))
           .bichain(
@@ -83,6 +74,74 @@ export function deployContractWith ({ fetch, WARP_GATEWAY_URL, logger: _logger, 
            * for shape
            */
           .map(res => ({ res, contractId: res.contractTxId }))
+      )
+      .toPromise()
+  }
+}
+
+export function registerContractWith ({ fetch, WARP_GATEWAY_URL, IRYS_NODE, logger: _logger }) {
+  const logger = _logger.child('warp-gateway:registerContract')
+
+  const irysNode = z.enum(['node1', 'node2'], {
+    errorMap: () => ({ message: 'the bundlr/irys node must be \'node1\' or \'node2\'' })
+  }).parse(IRYS_NODE)
+
+  return (args) => {
+    return of(args)
+      .map(logger.tap('registering bundled contract with the warp gateway %s', `${WARP_GATEWAY_URL}/gateway/v2/contracts/register`))
+      .chain(({ contractId }) =>
+        of(contractId)
+          .chain(fromPromise(async (contractId) =>
+            fetch(
+              `${WARP_GATEWAY_URL}/gateway/contracts/register`,
+              {
+                method: 'POST',
+                headers: {
+                  'Content-Type': 'application/json',
+                  Accept: 'application/json'
+                },
+                body: JSON.stringify({
+                  /**
+                   * Docs say contractId, but impl uses id, so we'll just include both
+                   * to hedge
+                   */
+                  contractId,
+                  id: contractId,
+                  /**
+                   * Warp requires passing the bundlr node that was used to upload
+                   * the contract. But Bundlr recently rebranded to Irys.
+                   *
+                   * So we are hedging for whenever warp changes public to match Bundlr's
+                   * rebranding to Irys.
+                   *
+                   */
+                  irysNode,
+                  bundlrNode: irysNode,
+                  node: irysNode
+                })
+              }
+            )
+          ))
+          .bichain(
+            err => Rejected(new Error(`Error while communicating with warp gateway: ${JSON.stringify(err)}`)),
+            fromPromise(
+              async res => {
+                if (res.ok) return res.json()
+                throw new Error(`${res.status}: ${await res.text()}`)
+              }
+            )
+          )
+          .bimap(
+            logger.tap('Error encountered when registering bundled contract with the warp gateway'),
+            logger.tap('Successfully registered bundled contract with the warp gateway')
+          )
+          /**
+           * See https://github.com/warp-contracts/gateway/blob/a1192869de24426a973465cf5be0a37b27a4c5ff/src/gateway/router/routes/registerContractRoute.ts#L124
+           * for shape
+           *
+           * ignore the returned contractTxId and just return the contractId provided
+           */
+          .map(res => ({ res, contractId }))
       )
       .toPromise()
   }