+++ sdk/test/browser-demo/src/app.js
@@ -1,0 +1,6 @@
+import App from './App.svelte'
+
+globalThis.localStorage.setItem('debug', '@permaweb/ao-sdk*')
+
+// eslint-disable-next-line
+new App({ target: document.body })