+++ servers/cu/src/domain/lib/readState/evaluate.js
@@ -1,4 +1,8 @@
-import { __, assoc, assocPath, prop, reduce, reduced } from 'ramda'
+import {
+  T,
+  __, always, applySpec, assoc, assocPath, concat, cond, endsWith, identity,
+  ifElse, is, isNotNil, path, pathOr, pipe, propOr, reduce, reduced
+} from 'ramda'
 import { fromPromise, of, Rejected, Resolved } from 'hyper-async'
 import AoLoader from '@permaweb/ao-loader'
 import { z } from 'zod'
@@ -54,7 +58,22 @@ function cacheEvaluationWith ({ saveEvaluation, logger }) {
 export function evaluateWith (env) {
   const logger = env.logger.child('evaluate')
 
+  const ACCUMULATE_RESULT = env.ACCUMULATE_RESULT
+
   const cacheEvaluation = cacheEvaluationWith({ ...env, logger })
+
+  /**
+   * Whether we need to accumulate certain values during evaluation
+   * is not clear. So we are using an environment variable to toggle
+   * certain functionality related to accumulation.
+   *
+   * TODO: remove once we find out.
+   */
+  const maybeAccumulateResult = (fn) => ifElse(
+    always(ACCUMULATE_RESULT),
+    fn,
+    identity
+  )
 
   /**
    * When an error occurs, we short circuit the reduce using
@@ -77,6 +96,75 @@ export function evaluateWith (env) {
   const maybeResolveError = maybeReducedError(Resolved)
   const maybeRejectError = maybeReducedError(Rejected)
 
+  /**
+   * Given the previous interaction output,
+   * return a function that will merge the next interaction output
+   * with the previous.
+   */
+  const mergeOutput = (prev) => applySpec({
+    /**
+     * We default to 'new' state, from applying an interaction,
+     * to the previous state, but it will be overwritten by the outputs state
+     *
+     * This ensures the new interaction in the chain has state to
+     * operate on, even if the previous interaction only produced
+     * messages and no state change.
+     */
+    state: propOr(prev.state, 'state'),
+    result: {
+      /**
+       * If an interaction produces an error simply pass it through
+       */
+      error: path(['result', 'error']),
+      /**
+       * It is up the consumer ie. an mu, to keep track of which messages
+       * it has already processed. The cu's responsibility is to just evaluate
+       * deterministically for a given range of sequenced interactions
+       */
+      messages: pipe(
+        pathOr([], ['result', 'messages']),
+        maybeAccumulateResult(concat(prev.result.messages))
+      ),
+      /**
+       * It is up the consumer ie. an mu, to keep track of which spawns
+       * it has already processed. The cu's responsibility is to just evaluate
+       * deterministically for a given range of sequenced interactions
+       */
+      spawns: pipe(
+        pathOr([], ['result', 'spawns']),
+        maybeAccumulateResult(concat(prev.result.spawns))
+      ),
+      output: pipe(
+        path(['result', 'output']),
+        ifElse(
+          isNotNil,
+          pipe(
+            /**
+             * Always make sure the output
+             * is a string
+             */
+            cond([
+              [is(String), identity],
+              [is(Number), String],
+              [is(Object), obj => JSON.stringify(obj)],
+              [T, identity]
+            ]),
+            /**
+             * Ensure the output from the interaction ends with a newline
+             */
+            ifElse(
+              endsWith('\n'),
+              identity,
+              maybeAccumulateResult(concat(__, '\n'))
+            ),
+            maybeAccumulateResult(concat(prev.result.output))
+          ),
+          always(prev.result.output)
+        )
+      )
+    }
+  })
+
   return (ctx) =>
     of(ctx)
       .chain(addHandler)
@@ -88,9 +176,8 @@ export function evaluateWith (env) {
           ($output, { action, sortKey, SWGlobal }) =>
             $output
               .chain(maybeRejectError)
-              .map(prop('state'))
-              .chain((state) =>
-                of(state)
+              .chain((prev) =>
+                of(prev.state)
                   .chain(
                     fromPromise((state) => ctx.handle(state, action, SWGlobal))
                   )
@@ -101,20 +188,15 @@ export function evaluateWith (env) {
                     (err) => Resolved(assocPath(['result', 'error'], err, {})),
                     Resolved
                   )
+                  /**
+                   * The the previous interaction output, and merge it
+                   * with the output of the current interaction
+                   */
+                  .map(mergeOutput(prev))
                   .chain((output) => {
-                    if (output.result && output.result.error) {
-                      return Rejected(output)
-                    }
-                    /**
-                     * We default to state to the previous state,
-                     * but it will be overwritten by the spread
-                     * if output contains state.
-                     *
-                     * This ensures the new interaction in the chain has state to
-                     * operate on, even if the previous interaction only produced
-                     * messages and no state change.
-                     */
-                    return Resolved({ state, ...output })
+                    return output.result && output.result.error
+                      ? Rejected(output)
+                      : Resolved(output)
                   })
               )
               .bimap(
@@ -147,7 +229,14 @@ export function evaluateWith (env) {
                  */
                 Resolved
               ),
-          of({ state: ctx.state, result: ctx.result }),
+          of({
+            state: ctx.state,
+            /**
+             * Every evaluation range starts with no messages, no output,
+             * and no spawns
+             */
+            result: { messages: [], output: '', spawns: [] }
+          }),
           ctx.actions
         )
       )