+++ servers/cu/src/domain/lib/readState/evaluate.test.js
@@ -8,13 +8,12 @@ import { evaluateWith } from './evaluate.js'
 const logger = createLogger('ao-cu:readState')
 
 describe('evaluate', () => {
-  test('evaluate state and add output to context', async () => {
-    const env = {
+  describe('output - ACCUMULATE_RESULT', () => {
+    const evaluate = evaluateWith({
+      ACCUMULATE_RESULT: true,
       saveEvaluation: async (interaction) => interaction,
       logger
-    }
-
-    const evaluate = evaluateWith(env)
+    })
 
     const ctx = {
       id: 'ctr-1234',
@@ -35,11 +34,112 @@ describe('evaluate', () => {
       ]
     }
 
-    const res = await evaluate(ctx).toPromise()
-    console.log(res)
-    assert.ok(res.output)
-    assert.deepStrictEqual(res.output, {
-      state: { heardHello: true, heardWorld: true, happy: true }
+    test('adds output to context', async () => {
+      const { output } = await evaluate(ctx).toPromise()
+      assert.ok(output)
+    })
+
+    test('folds the state', async () => {
+      const { output } = await evaluate(ctx).toPromise()
+      assert.deepStrictEqual(output.state, { heardHello: true, heardWorld: true, happy: true })
+    })
+
+    test('accumulates the result.messages', async () => {
+      const expectedMessage = {
+        target: 'contract-foo-123',
+        input: { function: 'noop' },
+        tags: [
+          { name: 'foo', value: 'bar' }
+        ]
+      }
+      const { output } = await evaluate(ctx).toPromise()
+      assert.deepStrictEqual(output.result.messages, [expectedMessage, expectedMessage])
+    })
+
+    test('accumulates the result.spawns', async () => {
+      const expectedSpawn = {
+        src: 'contract-src-123',
+        initState: { balances: { foo: 0 } },
+        tags: [
+          { name: 'foo', value: 'bar' }
+        ]
+      }
+      const { output } = await evaluate(ctx).toPromise()
+      assert.deepStrictEqual(output.result.spawns, [expectedSpawn, expectedSpawn])
+    })
+
+    test('accumulates the result.output', async () => {
+      const { output } = await evaluate(ctx).toPromise()
+      assert.deepStrictEqual(output.result.output, 'foobar\nfoobar\n')
+    })
+  })
+
+  describe('output - NOT ACCUMULATE_RESULT', () => {
+    const evaluate = evaluateWith({
+      /**
+       * disable accumulation of the result
+       */
+      ACCUMULATE_RESULT: false,
+      saveEvaluation: async (interaction) => interaction,
+      logger
+    })
+
+    const ctx = {
+      id: 'ctr-1234',
+      from: 'sort-key-start',
+      src: readFileSync('./test/contracts/happy/contract.wasm'),
+      state: {},
+      actions: [
+        {
+          action: { input: { function: 'hello' } },
+          sortKey: 'a',
+          SWGlobal: {}
+        },
+        {
+          action: { input: { function: 'world' } },
+          sortKey: 'b',
+          SWGlobal: {}
+        }
+      ]
+    }
+
+    test('adds output to context', async () => {
+      const { output } = await evaluate(ctx).toPromise()
+      assert.ok(output)
+    })
+
+    test('folds the state', async () => {
+      const { output } = await evaluate(ctx).toPromise()
+      assert.deepStrictEqual(output.state, { heardHello: true, heardWorld: true, happy: true })
+    })
+
+    test('DOES NOT accumulate the result.messages', async () => {
+      const expectedMessage = {
+        target: 'contract-foo-123',
+        input: { function: 'noop' },
+        tags: [
+          { name: 'foo', value: 'bar' }
+        ]
+      }
+      const { output } = await evaluate(ctx).toPromise()
+      assert.deepStrictEqual(output.result.messages, [expectedMessage])
+    })
+
+    test('DOES NOT accumulate the result.spawns', async () => {
+      const expectedSpawn = {
+        src: 'contract-src-123',
+        initState: { balances: { foo: 0 } },
+        tags: [
+          { name: 'foo', value: 'bar' }
+        ]
+      }
+      const { output } = await evaluate(ctx).toPromise()
+      assert.deepStrictEqual(output.result.spawns, [expectedSpawn])
+    })
+
+    test('DOES NOT accumulates the result.output', async () => {
+      const { output } = await evaluate(ctx).toPromise()
+      assert.deepStrictEqual(output.result.output, 'foobar')
     })
   })
 
@@ -112,7 +212,12 @@ describe('evaluate', () => {
 
     const { output } = await evaluate(ctx).toPromise()
     assert.deepStrictEqual(output, {
-      state: { balances: { 1: 1 } }
+      state: { balances: { 1: 1 } },
+      result: {
+        messages: [],
+        spawns: [],
+        output: ''
+      }
     })
   })
 
@@ -142,8 +247,12 @@ describe('evaluate', () => {
     const res = await evaluate(ctx).toPromise()
     assert.ok(res.output)
     assert.deepStrictEqual(res.output, {
+      state: {},
       result: {
-        error: { code: 123, message: 'a handled error within the contract' }
+        error: { code: 123, message: 'a handled error within the contract' },
+        messages: [],
+        spawns: [],
+        output: ''
       }
     })
   })
@@ -174,8 +283,12 @@ describe('evaluate', () => {
     const res = await evaluate(ctx).toPromise()
     assert.ok(res.output)
     assert.deepStrictEqual(res.output, {
+      state: {},
       result: {
-        error: { code: 123, message: 'a thrown error within the contract' }
+        error: { code: 123, message: 'a thrown error within the contract' },
+        messages: [],
+        spawns: [],
+        output: ''
       }
     })
   })